\name{fastcode-package}
\alias{fastcode-package}
\alias{fastcode}
\docType{package}
\title{
\packageTitle{fastcode}
}
\description{
\packageDescription{fastcode}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastcode}
\packageIndices{fastcode}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fastcode}

Maintainer: \packageMaintainer{fastcode}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
