% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A matrix with name v and w.}

\item{W}{A number.}
}
\value{
A list with value and elements.
}
\description{
Greedy heuristic
}
\examples{
library("parallel")
data(knapsack_objects)
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}

