% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fast.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{The only solution that is guaranteed to give a correct answer in all
situations for the knapsack problem
is using brute-force search, i.e. going through all possible alternatives
and return the maximum value
found. This approach is of complexity O(2 n ) since all possible
combinations 2 n needs to be evaluated.}
\usage{
knapsack_brute_force(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A matrix with name v and w.}

\item{W}{A number.}

\item{parallel}{Multicore control.}
}
\value{
A list with value and elements.
}
\description{
The only solution that is guaranteed to give a correct answer in all
situations for the knapsack problem
is using brute-force search, i.e. going through all possible alternatives
and return the maximum value
found. This approach is of complexity O(2 n ) since all possible
combinations 2 n needs to be evaluated.
}
\examples{
data(knapsack_objects)
knapsack_brute_force(x = knapsack_objects[1:8,], W = 3500)
}

