% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{We will now take another approach to the problem. If the weights are actually
discrete values (as in our example) we can use this to create an algorithm
that can solve the knapsack problem exact by iterating
over all possible values of w.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A matrix with name v and w.}

\item{W}{A number.}
}
\value{
A list with value and elements.
}
\description{
We will now take another approach to the problem. If the weights are actually
discrete values (as in our example) we can use this to create an algorithm
that can solve the knapsack problem exact by iterating
over all possible values of w.
}
\examples{
data(knapsack_objects)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
}

