% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{dataset
for test the function}
\format{a data frame with 2000 rows and 2 variables

\describe{
\item{w}{weight}
\item{v}{value}
}}
\usage{
knapsack_objects
}
\description{
dataset
for test the function
}
\keyword{datasets}

